+++ tests/Integration/testData/Valid/.ide-helper.php
@@ -5,19 +5,19 @@
 
 namespace {
 
-    /** @deprecated since v6.0.0. This class will be remove in the future. Please use the corresponding class from the unified namespace. */
+    /** @deprecated since v6.0.0. This class will be removed in the future. Please use the corresponding class from the unified namespace. */
     class oxClassExistsInCommunityEdition extends \OxidEsales\Eshop\ClassExistsInCommunityEdition
     {
 
     }
 
-    /** @deprecated since v6.0.0. This class will be remove in the future. Please use the corresponding class from the unified namespace. */
+    /** @deprecated since v6.0.0. This class will be removed in the future. Please use the corresponding class from the unified namespace. */
     interface oxInterface extends \OxidEsales\Eshop\OxidInterface
     {
 
     }
 
-    /** @deprecated since v6.0.0. This class will be remove in the future. Please use the corresponding class from the unified namespace. */
+    /** @deprecated since v6.0.0. This class will be removed in the future. Please use the corresponding class from the unified namespace. */
     abstract class oxAbstractClass extends \OxidEsales\Eshop\OxidAbstractClass
     {